+++ runtime/core/src/routing/Wiring.kt
@@ -1,5 +1,6 @@
 package io.github.darvld.wireframe.routing
 
+import graphql.schema.GraphQLScalarType
 import io.github.darvld.wireframe.WireframeInternal
 import io.github.darvld.wireframe.execution.SuspendingDataFetcher
 
@@ -7,4 +8,7 @@ import io.github.darvld.wireframe.execution.SuspendingDataFetcher
 internal sealed interface Wiring {
     @JvmInline
     value class DataFetcherWiring(val fetcher: SuspendingDataFetcher<Any?>) : Wiring
+
+    @JvmInline
+    value class ScalarWiring(val scalar: GraphQLScalarType) : Wiring
 }